api/requirements.txt
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
decorator==4.3.0
Django==2.1.2
djangorestframework==3.8.2
future==0.17.1
idna==2.8
image==1.5.27
imutil==0.1.22
imutils==0.5.2
kiwisolver==1.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.4
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
Pillow==5.4.1
pip==10.0.1
pylablib==0.3.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.5
PyWavelets==1.0.1
requests==2.21.0
scikit-image==0.14.1
scikit-learn==0.20.2
scipy==1.2.0
setuptools==39.1.0
six==1.12.0
sklearn==0.0
todo==0.1
toolz==0.9.0
urllib3==1.24.1
